move survey validation from model to request
