+++ src/Ushahidi/Modules/V5/Requests/UserRequest.php
@@ -44,6 +44,11 @@ class UserRequest extends BaseRequest
         }
     }
 
+    /**
+     * Get the error messages for the defined validation rules.
+     *
+     * @return array
+     */
     public function messages()
     {
         return [
